+++ NEWS.md
@@ -1,6 +1,7 @@
 # err 1.0.1
 
 - Add `latche`, and `latchw` to latche errors and warnings onto object to later `enforce` them.
+- Added `defer_enforce` to defer the enforce to when the function exits.
 
 # err 1.0.0